backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ML / AI
tensorflow==2.18.0
keras==3.9.0

# Data Processing
numpy==1.26.4
pillow==11.1.0

# Web & Forms
python-multipart==0.0.20

# Security
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.10.5
pydantic-settings==2.7.1

# Database
sqlmodel==0.0.14

# Logging
loguru==0.7.3

# Utilities
python-dotenv==1.0.1
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0